move Slack class to separate file
